Call onHeightChange instead of re-mounting cards
